jax_lm/pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0", "hatch-fancy-pypi-readme>=24.1.0"]
build-backend = "hatchling.build"

[project]
name = "jax-diffucoder"
version = "0.1.0"
description = "High-performance JAX/Flax implementation of DiffuCoder for code generation"
readme = "README.md"
authors = [
    {name = "ML-DiffuCoder Contributors", email = "info@atsentia.com"}
]
license = {text = "Apache-2.0"}
keywords = [
    "jax",
    "flax", 
    "diffusion",
    "code-generation",
    "machine-learning",
    "tpu",
    "language-model"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"

dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "flax>=0.7.5",
    "optax>=0.1.7",
    "orbax-checkpoint>=0.4.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "tensorstore>=0.1.45",
    "ml-collections>=0.1.1",
    "absl-py>=1.4.0",
]

[project.optional-dependencies]
gpu = [
    "jax[cuda12_pip]>=0.4.20",
]
tpu = [
    "jax[tpu]>=0.4.20",
]
dev = [
    "pytest>=7.3.0",
    "pytest-xdist>=3.3.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]
tokenizer = [
    "tokenizers>=0.15.0",
    "regex>=2023.0.0",
]
# Minimal HF dependency only for file downloads
hf = [
    "huggingface-hub>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/atsentia/ml-diffucoder"
Repository = "https://github.com/atsentia/ml-diffucoder"
Documentation = "https://github.com/atsentia/ml-diffucoder/tree/main/jax_lm"
"Bug Tracker" = "https://github.com/atsentia/ml-diffucoder/issues"

[project.scripts]
jax-diffucoder = "jax_lm.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "/jax_lm",
    "/tests",
    "/README.md",
    "/LICENSE",
]
exclude = [
    "*.pyc",
    "__pycache__",
    ".git",
    ".uv_cache",
    "build",
    "dist",
]

[tool.hatch.build.targets.wheel]
packages = ["jax_lm"]

[tool.hatch.build.targets.wheel.shared-data]
"jax_lm/py.typed" = "jax_lm/py.typed"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true